Create Bilinear distribution for initializing Transposed Convolution kernels
Let's elaborate on the layer type used for upsampling in an FCN that we learned about in the Theory chapter: transposed convolution. Recall that the kernel weights act as a kind of "brush" distributor of an input pixel's value, and we found a good setting for overlapping sliding window positions with
𝜃(𝑥)=(𝑠,𝑘,𝑝)𝑇=(𝑥,2𝑥,𝑥/2)𝑇,
 
where  𝑥  is the factor we want to upsample by, and  𝑠 ,  𝑘 ,  𝑝  are the stride, kernel size, and padding parameters of the transposed convolution layer, respectively.

According to the brush analogy, if we set all the weights within a transposed convolution kernel to the same value, we would achieve a drawing pattern similar to a marker pen: We would color all output pixels in the sliding window with the same intensity, which would be derived from the corresponding input pixel.

From the fully convolutional MobileNet base topology, we found out that we downsample by a factor of 32 horizontally and vertically. Correspondingly, in order to reach a classification granularity of one distinct class per pixel, we would have to eventually upsample by exactly the same factor. This means we can use the configuration for  𝑠 ,  𝑘 , and  𝑝  that we calculated earlier.

This implies that we have rather large square kernels with an edge length of  𝑘=64 ; instead of using a marker pen, we would actually prefer a "real" brush shape where we paint more color in central regions than in the surrounding area. If you are familiar with image processing libraries such as OpenCV or OpenVX, this strategy is often referred to as upsampling with bilinear interpolation (Prashanth et al., 2009), and it can also be used for FCNs as well Long et al. (2015).

Within the terminology of transposed convolution, we could interpret this as a setup where the pixels within a moving window on the output feature map that are further away from the central input pixel contribute less to the final prediction than those that are spacially closer.