Create A complete FCN head for Semantic Segmentation with a MobileNet stem
In this section, you'll put together a number of concepts from earlier in the notebook. Read carefully and check yourself to be sure you have understood everything.

We will construct a very simple Fully Convolutional Network (Long et al., 2015) head to our MobileNet stem that enables training the overall architecture for a semantic segmentation task. Even though this FCN head is very basic – do not expect to win any competitions with it – it will do an adequate job on our Cityscapes dataset and can serve as solid learning material.

Here is the architecture of mentioned FCN head that we will subsequently implement (note that we have omitted the activation and batch normalization layers for all convolution layers):


As you can see, we append it to a pruned MobileNet with a spatial output dimensionality of  10𝑥15 , whereas the depth of the MobileNet output is determined by 1024*d_mul (making use of the Depth Multiplier parameter, here d_mul, that we previously described). In the code below, d_mul is referred to as depth_multiplier.

The head's first convolution layer is there to replace the average pooling layer of the classification MobileNet topology with a window size of  7𝑥7  (remember your findings from Exercise A). For this layer, we have to choose a kernel size k, while we use TensorFlow's 'FULL' padding policy and a stride of 1 so that our resolution does not change. The greater the value for k, the greater the receptive field of the network, which basically represents the size of an area on the input image that is responsible for the prediction of one output pixel. To keep the computational needs within that layer low, we use the depth-wise convolution concept that is prominently present in MobileNets. Correspondingly, we first have the separable convolution layer conv1_separable, producing 1024*d_mul output feature maps, which is followed by a point-wise convolution layer, which we will describe in the following bullet point. In the code below, k is referred to as kernel.

The point-wise convolution layer conv1_pointwise has a  1𝑥1  kernel with d filter maps, so that the output feature map depth is equal to d. In the code below, d is referred to as depth.
In the subsequent layer conv2_classifier, we intend to classify each of the  10𝑥15  pixels received from the previous layer into n classes. Therefore, we use another  1𝑥1  convolution layer with n filters. In the code below, n is referred to as num_classes.
In principle, the output feature maps from the previous  1𝑥1  convolution layer could already serve as a very coarse-grained semantic segmentation task, where a region of  32𝑥32  input pixels would be processed to a prediction of a single pixel. Our goal is to output predictions at a (single-)pixel granularity; we will therefore make use of the transposed convolution layer upsampler with the parameters from the question you answered earlier, and initialize the weights with the function we explored in Exercise C. All the previous layers are initialized with the Xavier weight filler.